+++ OgreMain/include/OgreMurmurHash3.h
@@ -9,6 +9,8 @@
 // Platform-specific functions and macros
 #include "OgrePlatform.h"
 
+#include <cstddef>
+
 // Microsoft Visual Studio
 
 #if defined( _MSC_VER ) && _MSC_VER < 1600
@@ -31,13 +33,13 @@ namespace Ogre
 
 namespace Ogre
 {
-    void _OgreExport MurmurHash3_x86_32  ( const void * key, int len, uint32_t seed, void * out );
+    void _OgreExport MurmurHash3_x86_32  ( const void * key, size_t len, uint32_t seed, void * out );
 
-    void _OgreExport MurmurHash3_x86_128 ( const void * key, int len, uint32_t seed, void * out );
+    void _OgreExport MurmurHash3_x86_128 ( const void * key, size_t len, uint32_t seed, void * out );
 
-    void _OgreExport MurmurHash3_x64_128 ( const void * key, int len, uint32_t seed, void * out );
+    void _OgreExport MurmurHash3_x64_128 ( const void * key, size_t len, uint32_t seed, void * out );
 
-    inline void MurmurHash3_128( const void * key, int len, uint32_t seed, void * out ) {
+    inline void MurmurHash3_128( const void * key, size_t len, uint32_t seed, void * out ) {
 #if OGRE_ARCH_TYPE == OGRE_ARCHITECTURE_64
         MurmurHash3_x64_128(key, len, seed, out);
 #else
